requirements.txt
# Django Backend Dependencies
annotated-types==0.7.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
Faker==37.5.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0

# FastAPI Recommender Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# ML & Data Science Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
node2vec>=0.4.6
networkx>=3.1
scipy>=1.11.0
matplotlib>=3.7.0
joblib>=1.3.0
pydantic>=2.0.0

# Database
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic_core==2.33.2

# Image Processing
pillow==11.3.0

# Additional Dependencies
proto-plus==1.26.1
protobuf==5.29.5
